+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/guide/options/SlimefunGuideSettings.java
@@ -47,6 +47,7 @@ public final class SlimefunGuideSettings {
     static {
         options.add(new GuideModeOption());
         options.add(new FireworksOption());
+        options.add(new LearningAnimationOption());
         options.add(new PlayerLanguageOption());
     }
 
@@ -238,15 +239,51 @@ public final class SlimefunGuideSettings {
      * @return Whether this {@link Player} wants to see fireworks when unlocking a {@link Research}
      */
     public static boolean hasFireworksEnabled(@Nonnull Player p) {
+        return getOptionValue(p, FireworksOption.class, true);
+    }
+
+    /**
+     * This method checks if the given {@link Player} has enabled the {@link LearningAnimationOption}
+     * in their {@link SlimefunGuide}.
+     * If they enabled this setting, they will see messages in chat about the progress of their {@link Research}.
+     *
+     * @param p
+     *            The {@link Player}
+     *
+     * @return Whether this {@link Player} wants to info messages in chat when unlocking a {@link Research}
+     */
+    public static boolean hasLearningAnimationEnabled(@Nonnull Player p) {
+        return getOptionValue(p, LearningAnimationOption.class, true);
+    }
+
+    /**
+     * Helper method to get the value of a {@link SlimefunGuideOption} that the {@link Player}
+     * has set in their {@link SlimefunGuide}
+     *
+     * @param p
+     *            The {@link Player}
+     * @param optionsClass
+     *            Class of the {@link SlimefunGuideOption} to get the value of
+     * @param defaultValue
+     *            Default value to return in case the option is not found at all or has no value set
+     * @param <T>
+     *            Type of the {@link SlimefunGuideOption}
+     * @param <V>
+     *            Type of the {@link SlimefunGuideOption} value
+     * 
+     * @return The value of given {@link SlimefunGuideOption}
+     */
+    @Nonnull
+    private static <T extends SlimefunGuideOption<V>, V> V getOptionValue(@Nonnull Player p, @Nonnull Class<T> optionsClass, @Nonnull V defaultValue) {
         for (SlimefunGuideOption<?> option : options) {
-            if (option instanceof FireworksOption) {
-                FireworksOption fireworks = (FireworksOption) option;
+            if (optionsClass.isInstance(option)) {
+                T o = optionsClass.cast(option);
                 ItemStack guide = SlimefunGuide.getItem(SlimefunGuideMode.SURVIVAL_MODE);
-                return fireworks.getSelectedOption(p, guide).orElse(true);
+                return o.getSelectedOption(p, guide).orElse(defaultValue);
             }
         }
 
-        return true;
+        return defaultValue;
     }
 
 }