+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/enchanting/AutoEnchanter.java
@@ -1,30 +1,24 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.electric.machines.enchanting;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.annotation.ParametersAreNonnullByDefault;
-
+import io.github.thebusybiscuit.cscorelib2.inventory.InvUtils;
+import io.github.thebusybiscuit.slimefun4.api.events.AsyncAutoEnchanterProcessEvent;
 import io.github.thebusybiscuit.slimefun4.api.events.AutoEnchantEvent;
+import me.mrCookieSlime.Slimefun.Lists.RecipeType;
+import me.mrCookieSlime.Slimefun.Objects.Category;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.MachineRecipe;
+import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
+import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.EnchantmentStorageMeta;
 
-import io.github.thebusybiscuit.cscorelib2.chat.ChatColors;
-import io.github.thebusybiscuit.cscorelib2.inventory.InvUtils;
-import io.github.thebusybiscuit.cscorelib2.item.CustomItem;
-import io.github.thebusybiscuit.slimefun4.api.items.ItemSetting;
-import io.github.thebusybiscuit.slimefun4.api.items.settings.IntRangeSetting;
-import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
-import me.mrCookieSlime.Slimefun.Lists.RecipeType;
-import me.mrCookieSlime.Slimefun.Objects.Category;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.AContainer;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.MachineRecipe;
-import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
-import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
+import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * The {@link AutoEnchanter}, in contrast to the {@link AutoDisenchanter}, adds
@@ -35,21 +29,16 @@ import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
  * @author Poslovitch
  * @author Mooy1
  * @author StarWishSama
+ * @author martinbrom
  *
  * @see AutoDisenchanter
  *
  */
-public class AutoEnchanter extends AContainer {
-
-    private final ItemSetting<Boolean> useEnchantLevelLimit = new ItemSetting<>(this, "use-enchant-level-limit", false);
-    private final IntRangeSetting enchantLevelLimit = new IntRangeSetting(this, "enchant-level-limit", 0, 10, Short.MAX_VALUE);
+public class AutoEnchanter extends AbstractEnchantmentMachine {
 
     @ParametersAreNonnullByDefault
     public AutoEnchanter(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
-
-        addItemSetting(useEnchantLevelLimit);
-        addItemSetting(enchantLevelLimit);
     }
 
     @Override
@@ -60,79 +49,87 @@ public class AutoEnchanter extends AContainer {
     @Override
     protected MachineRecipe findNextRecipe(BlockMenu menu) {
         for (int slot : getInputSlots()) {
-            ItemStack target = menu.getItemInSlot(slot == getInputSlots()[0] ? getInputSlots()[1] : getInputSlots()[0]);
+            ItemStack item = menu.getItemInSlot(slot == getInputSlots()[0] ? getInputSlots()[1] : getInputSlots()[0]);
 
             // Check if the item is enchantable
-            if (!isEnchantable(target)) {
-                return null;
+            if (!isEnchantable(item)) {
+                continue;
             }
 
-            AutoEnchantEvent event = new AutoEnchantEvent(target);
+            // Call an event so other Plugins can modify it.
+            AutoEnchantEvent event = new AutoEnchantEvent(item);
             Bukkit.getPluginManager().callEvent(event);
 
             if (event.isCancelled()) {
                 return null;
             }
 
-            ItemStack item = menu.getItemInSlot(slot);
+            ItemStack enchantedBook = menu.getItemInSlot(slot);
 
-            if (item != null && item.getType() == Material.ENCHANTED_BOOK && target != null) {
-                Map<Enchantment, Integer> enchantments = new HashMap<>();
-                int amount = 0;
-                EnchantmentStorageMeta meta = (EnchantmentStorageMeta) item.getItemMeta();
-
-                for (Map.Entry<Enchantment, Integer> e : meta.getStoredEnchants().entrySet()) {
-                    if (e.getKey().canEnchantItem(target)) {
-                        if (!useEnchantLevelLimit.getValue() || enchantLevelLimit.getValue() >= e.getValue()) {
-                            amount++;
-                            enchantments.put(e.getKey(), e.getValue());
-                        } else if (!menu.toInventory().getViewers().isEmpty()) {
-                            String notice = ChatColors.color(SlimefunPlugin.getLocalization().getMessage("messages.above-limit-level"));
-                            notice = notice.replace("%level%", String.valueOf(enchantLevelLimit.getValue()));
-                            ItemStack progressBar = new CustomItem(Material.BARRIER, " ", notice);
-                            menu.replaceExistingItem(22, progressBar);
-                            return null;
-                        }
-                    }
-                }
-
-                if (amount > 0) {
-                    ItemStack enchantedItem = target.clone();
-                    enchantedItem.setAmount(1);
-
-                    for (Map.Entry<Enchantment, Integer> entry : enchantments.entrySet()) {
-                        enchantedItem.addUnsafeEnchantment(entry.getKey(), entry.getValue());
-                    }
-
-                    MachineRecipe recipe = new MachineRecipe(75 * amount / this.getSpeed(), new ItemStack[] { target, item }, new ItemStack[] { enchantedItem, new ItemStack(Material.BOOK) });
-
-                    if (!InvUtils.fitAll(menu.toInventory(), recipe.getOutput(), getOutputSlots())) {
-                        return null;
-                    }
-
-                    for (int inputSlot : getInputSlots()) {
-                        menu.consumeItem(inputSlot);
-                    }
-
-                    return recipe;
-                }
-
-                return null;
+            if (enchantedBook != null && enchantedBook.getType() == Material.ENCHANTED_BOOK) {
+                return enchant(menu, item, enchantedBook);
             }
         }
 
         return null;
     }
 
-    private boolean isEnchantable(ItemStack item) {
-        SlimefunItem sfItem = null;
+    @Nullable
+    @ParametersAreNonnullByDefault
+    protected MachineRecipe enchant(BlockMenu menu, ItemStack target, ItemStack enchantedBook) {
+        // Call an event so other Plugins can modify it.
+        AsyncAutoEnchanterProcessEvent event = new AsyncAutoEnchanterProcessEvent(target, enchantedBook, menu);
+        Bukkit.getPluginManager().callEvent(event);
 
-        // stops endless checks of getByItem for enchanted book stacks.
-        if (item != null && item.getType() != Material.ENCHANTED_BOOK) {
-            sfItem = SlimefunItem.getByItem(item);
+        if (event.isCancelled()) {
+            return null;
         }
 
-        return sfItem == null || sfItem.isEnchantable();
+        EnchantmentStorageMeta meta = (EnchantmentStorageMeta) enchantedBook.getItemMeta();
+        Map<Enchantment, Integer> enchantments = new HashMap<>();
+
+        // Find applicable enchantments
+        for (Map.Entry<Enchantment, Integer> entry : meta.getStoredEnchants().entrySet()) {
+            if (entry.getKey().canEnchantItem(target)) {
+                if (isEnchantmentLevelAllowed(entry.getValue())) {
+                    enchantments.put(entry.getKey(), entry.getValue());
+                } else if (!menu.toInventory().getViewers().isEmpty()) {
+                    showEnchantmentLevelWarning(menu);
+                    return null;
+                }
+            }
+        }
+
+        // Check if we found any valid enchantments
+        if (!enchantments.isEmpty()) {
+            ItemStack enchantedItem = target.clone();
+            enchantedItem.setAmount(1);
+            enchantedItem.addUnsafeEnchantments(enchantments);
+
+            MachineRecipe recipe = new MachineRecipe(75 * enchantments.size() / getSpeed(), new ItemStack[] { target, enchantedBook }, new ItemStack[] { enchantedItem, new ItemStack(Material.BOOK) });
+
+            if (!InvUtils.fitAll(menu.toInventory(), recipe.getOutput(), getOutputSlots())) {
+                return null;
+            }
+
+            for (int inputSlot : getInputSlots()) {
+                menu.consumeItem(inputSlot);
+            }
+
+            return recipe;
+        } else {
+            return null;
+        }
+    }
+
+    private boolean isEnchantable(@Nullable ItemStack item) {
+        // stops endless checks of getByItem for enchanted book stacks.
+        if (item != null && item.getType() != Material.ENCHANTED_BOOK && !item.getType().isAir() && !hasIgnoredLore(item)) {
+            SlimefunItem sfItem = SlimefunItem.getByItem(item);
+            return sfItem == null || sfItem.isEnchantable();
+        } else {
+            return false;
+        }
     }
 
     @Override