+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/tools/ExplosivePickaxe.java
@@ -1,5 +1,7 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.tools;
 
+import javax.annotation.ParametersAreNonnullByDefault;
+
 import org.bukkit.block.Block;
 import org.bukkit.inventory.ItemStack;
 
@@ -19,6 +21,7 @@ import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
  */
 public class ExplosivePickaxe extends ExplosiveTool {
 
+    @ParametersAreNonnullByDefault
     public ExplosivePickaxe(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
     }